pyproject.toml
[project]
name = "clinic-appointment-system"
version = "0.1.0"
description = "Database Technology course final project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.5.0",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
]
